refactor(optimizer): cascade
