+++ src/query/sql/src/planner/optimizer/optimizers/cascades/mod.rs
@@ -13,11 +13,13 @@
 // limitations under the License.
 
 mod cascade;
+mod cost_model;
 mod explore_rules;
 mod scheduler;
 mod tasks;
 
 pub use cascade::CascadesOptimizer;
+pub use cost_model::DefaultCostModel;
 pub use scheduler::Scheduler;
 pub use scheduler::DEFAULT_TASK_LIMIT;
 pub use tasks::ExploreGroupTask;
